Rename to AzureDevOpsTestPublisher
