+++ src/Microsoft.DotNet.Helix/AzureDevOpsTestPublisher/Model/TerminalError.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.DotNet.Helix.AzureDevOpsTestReporter.Model;
+namespace Microsoft.DotNet.Helix.AzureDevOpsTestPublisher.Model;
 
 internal sealed class TerminalError(string message) : Exception(message)
 {
