+++ src/Microsoft.DotNet.Helix/JobMonitor/JobMonitorRunner.cs
@@ -14,7 +14,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Azure;
 using Azure.Storage.Blobs;
-using Microsoft.DotNet.Helix.AzureDevOpsTestReporter;
+using Microsoft.DotNet.Helix.AzureDevOpsTestPublisher;
 using Microsoft.DotNet.Helix.Client;
 using Microsoft.DotNet.Helix.Client.Models;
 using Newtonsoft.Json;
